feat: allowing infinite product stocks

+++ database/migrations/tenant/2023_06_20_033808_create_products_table.php
@@ -22,7 +22,7 @@ return new class () extends Migration {
             $table->decimal('selling_price', 10, 2);
             $table->json('dimension')->nullable();
             $table->decimal('weight')->nullable();
-            $table->unsignedInteger('stock')->nullable();
+            $table->unsignedBigInteger('stock')->nullable();
             $table->mediumText('description')->nullable();
             $table->boolean('status')->default(true);
             $table->boolean('is_digital_product')->default(false);
@@ -30,7 +30,7 @@ return new class () extends Migration {
             $table->boolean('is_special_offer')->default(false);
             $table->boolean('allow_customer_remarks')->default(false);
             $table->boolean('allow_stocks')->default(true);
-            $table->unsignedInteger('minimum_order_quantity')->default(1);
+            $table->unsignedBigInteger('minimum_order_quantity')->default(1);
             $table->timestamps();
         });
 
@@ -47,7 +47,7 @@ return new class () extends Migration {
             $table->json('combination');
             $table->decimal('retail_price', 10, 2);
             $table->decimal('selling_price', 10, 2);
-            $table->unsignedInteger('stock')->nullable();
+            $table->unsignedBigInteger('stock')->nullable();
             $table->boolean('status')->default(true);
             $table->timestamps();
         });
